requirements.txt
python-dotenv
langchain
langchain-chroma
langchain-google-genai
langchain-community
bs4
langchainhub
Flask